requirements.txt
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
boto3==1.28.2
botocore==1.31.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
deprecation==2.1.0
ecdsa==0.18.0
Flask==2.2.3
Flask-Cors==3.0.10
Flask-SQLAlchemy==3.0.3
frozenlist==1.3.3
greenlet==2.0.2
html2text==2020.1.16
idna==3.4
itsdangerous==2.1.2
jamo==0.4.1
Jinja2==3.1.2
jmespath==1.0.1
lxml==4.9.2
MarkupSafe==2.1.2
multidict==6.0.4
packaging==23.1
psycopg2-binary==2.9.6
pyasn1==0.4.8
PyJWT==2.7.0
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
python-jose==3.3.0
python-keycloak==2.15.3
pytz==2023.3
redis==4.5.5
requests==2.28.2
requests-toolbelt==0.10.1
rsa==4.9
s3transfer==0.6.1
six==1.16.0
SQLAlchemy==2.0.9
twilio==8.3.0
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.2.3
yarl==1.9.2
